Adjusting charths of dashboard
