+++ manager/src/helpers/formatters/date.ts
@@ -31,7 +31,7 @@ const formatInitialAndFinalDate = (initial: Date, final: Date) => {
 const formatToHumanDate = (date: string) => {
   const { dateFormat } = getCurrentLanguage();
 
-  return format(new Date(date).setHours(24), dateFormat);
+  return format(new Date(date), dateFormat);
 };
 
 export { formatInitialAndFinalDate, formatToHumanDate };
